Adds the ability to relaod output file to resume classification

+++ classification.py
@@ -2,6 +2,7 @@ from flask import session, request, jsonify, abort
 import tablib
 import json
 import csv
+from os import path
 
 in_data = tablib.Dataset()
 out_data = tablib.Dataset()
@@ -26,13 +27,25 @@ def load(opts):
         out_data.headers = [ header, "example", "nonexample", "unknown" ]
         fake_row = [ "", 0, 0, 0]
 
-        for row in range(len(in_data)):
-            out_data.append(fake_row)
-
         global out_data_path
         out_data_path = opts.output_data if opts.output_data else "out.csv"
-        with open(out_data_path, "w") as f:
-            f.write(out_data.csv)
+
+        if not path.exists(out_data_path):
+            if opts.debug: print("Create new output file.")
+            for row in range(len(in_data)):
+                out_data.append(fake_row)
+
+            with open(out_data_path, "w") as f:
+                f.write(out_data.csv)
+        else:
+            if opts.debug: print("Load existing output file.")
+            with open(out_data_path, "r") as f:
+                reader = csv.reader(f)
+                headers = next(reader)
+                out_data.headers = headers
+                for row in reader:
+                    out_data.append((row[0], int(row[1]), int(row[2]), int(row[3])))
+
 
 def register(app, opts):
